Adds solution for Functions exercise

+++ Exercises/Functions/ReturnTypes/decimal/decimal.ino
@@ -2,7 +2,7 @@
 // Hint: Look at the output from the Serial Monitor. Is it the same as the number you expected?
 // https://roboticsbackend.com/arduino-variable-types-complete-guide/
 
-int getOnePointFive()
+float getOnePointFive()
 {
     return 1.5;
 }
